% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cartesianProduct}
\alias{cartesianProduct}
\title{cartesianProduct}
\usage{
cartesianProduct(vec1, vec2)
}
\arguments{
\item{vec1}{first IntegerVector of values to compute the cartesian product}

\item{vec2}{second IntegerVector of values to compute the cartesian product}
}
\value{
IntegerMatrix: of 2 columns with the cartesian product of vec1 and vec2
}
\description{
cartesianProduct
}
