% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{compute_proba_control}
\alias{compute_proba_control}
\title{Proportion of a data set with a given variable at a given level}
\usage{
compute_proba_control(df, var, level)
}
\arguments{
\item{df}{Data frame.}

\item{var}{Character, column name in \code{df}.}

\item{level}{Value to match against \code{df[[var]]}.}
}
\value{
Numeric, proportion of rows where \code{df[[var]] == level} (na.rm=TRUE).
}
\description{
Proportion of a data set with a given variable at a given level
}
\examples{
df <- data.frame(colour = sample(c("orange", "purple"), 100, replace = TRUE))
compute_proba_control(df, "colour", "purple")
}
