% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Deltafind}
\alias{Deltafind}
\title{Deltafind()}
\usage{
Deltafind()
}
\value{
IntegerMatrix: find the indices of the elements in the cpp map _DeltaMap representing the sparse linkage matrix Delta.
}
\description{
Deltafind()
}
