% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ExpandGrid}
\alias{ExpandGrid}
\title{ExpandGrid}
\usage{
ExpandGrid(vector1, vector2)
}
\arguments{
\item{vector1}{first IntegerVector of values to compute the cartesian product}

\item{vector2}{second IntegerVector of values to compute the cartesian product}
}
\value{
IntegerMatrix: of 2 columns with the cartesian product of vec1 and vec2
}
\description{
ExpandGrid
}
