% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F2}
\alias{F2}
\title{F2}
\usage{
F2(U, nvals)
}
\arguments{
\item{U}{IntegerVector with factor values corresponding to the patterns of PIVs observed among records in the concerned source}

\item{nvals}{integer for the total number of possible patterns (among all sources)}
}
\value{
List: for each pattern in value, count of the records having the pattern in the concerned source
}
\description{
F2
}
