% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F33}
\alias{F33}
\title{F33}
\usage{
F33(A, B, nvals)
}
\arguments{
\item{A}{List with for each pattern in value, count of the records having the pattern in the concerned source}

\item{B}{List with for each pattern in value, count of the records having the pattern in the concerned source}

\item{nvals}{integer for the total number of possible patterns (among all sources)}
}
\value{
IntegerMatrix: nrow=nbr of potential links, ncol = 2, indicates the indices (of records from A, records from B) of records with matching patterns to be considered in the likelihood as potential links. It represents the cartesian product of both lists passed as parameters to represent all possible linked pairs of records.
}
\description{
F33
}
