% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{FDPinRL_BRL}
\alias{FDPinRL_BRL}
\title{FDP estimation with the BRL package}
\usage{
FDPinRL_BRL(NewA, NewB, arguments)
}
\arguments{
\item{NewA, }{NewB Data sets as returned by \code{\link[=synthesise]{synthesise()}}.}

\item{arguments}{List of extra arguments forwarded to \code{BRL::compareRecords()}
and \code{BRL::bipartiteGibbs()} (e.g. \code{flds}, \code{types}, \code{nIter}).}
}
\value{
List with \code{result_index_linked_A}, \code{result_index_linked_B}, \code{result_vec_linked_scores}.
}
\description{
More details about BRL on https://cran.r-project.org/web/packages/BRL/index.html
}
