% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateSequence}
\alias{generateSequence}
\title{generateSequence}
\usage{
generateSequence(n)
}
\arguments{
\item{n}{integer superior to 1}
}
\value{
IntegerVector: with n values from 1 to n
}
\description{
generateSequence
}
