% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initDeltaMap}
\alias{initDeltaMap}
\title{initDeltaMap}
\usage{
initDeltaMap()
}
\value{
void: Initialise the cpp map _DeltaMap representing the sparse linkage matrix Delta.
}
\description{
initDeltaMap
}
