% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{log_possible_config}
\alias{log_possible_config}
\title{Log number of possible linkage configurations}
\usage{
log_possible_config(n_records_B, sumD)
}
\arguments{
\item{n_records_B}{Integer, number of records in the larger data source (B).}

\item{sumD}{Integer, number of currently linked records.}
}
\value{
Numeric, \code{sum(log((n_records_B - sumD + 1):n_records_B))}, or \code{0} if
\code{sumD == 0}.
}
\description{
Computes \verb{log(nB! / (nB - sumD)!)}, i.e. the log number of ways to choose
\code{sumD} ordered links among \code{n_records_B} records of the larger file; used in
\code{\link[=log_lik]{log_lik()}} to normalise the likelihood of the linkage matrix.
}
\examples{
log_possible_config(n_records_B = 15, sumD = 5)
}
