% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{logPossibleConfig}
\alias{logPossibleConfig}
\title{Log number of possible linkage configurations}
\usage{
logPossibleConfig(Brecords, sumD)
}
\arguments{
\item{Brecords}{Integer, number of records in the larger data source (B).}

\item{sumD}{Integer, number of currently linked records.}
}
\value{
Numeric, \code{sum(log((Brecords - sumD + 1):Brecords))}, or \code{0} if \code{sumD == 0}.
}
\description{
Computes \verb{log(nB! / (nB - sumD)!)}, i.e. the log number of ways to choose
\code{sumD} ordered links among \code{Brecords} records of the larger file; used in
\code{\link[=loglik]{loglik()}} to normalise the likelihood of the linkage matrix.
}
\examples{
logPossibleConfig(Brecords = 15, sumD = 5)
}
