% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{plot.discrepancy_curves}
\alias{plot.discrepancy_curves}
\title{Plot post-linkage diagnostics over thresholds}
\usage{
\method{plot}{discrepancy_curves}(x, ...)
}
\arguments{
\item{x}{A \code{discrepancy_curves} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly; called for its plotting side effect.
}
\description{
Plot post-linkage diagnostics over thresholds
}
