% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{plot.FDP_curves}
\alias{plot.FDP_curves}
\title{Plot FDP estimates against the linkage decision threshold}
\usage{
\method{plot}{FDP_curves}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"FDP_curves"}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly; called for its plotting side effect.
}
\description{
Plot FDP estimates against the linkage decision threshold
}
