% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{print.RL_diagnostics}
\alias{print.RL_diagnostics}
\title{Print post-linkage diagnostics}
\usage{
\method{print}{RL_diagnostics}(x, threshold = 0.5, ...)
}
\arguments{
\item{x}{An \code{RL_diagnostics} object, see \code{\link[=RL_diagnostics]{RL_diagnostics()}}.}

\item{threshold}{Numeric, the linkage decision rule at which the measures
are displayed (default \code{0.5}).}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly; called for its printing side effect.
}
\description{
Print post-linkage diagnostics
}
