% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleH}
\alias{sampleH}
\title{sampleH}
\usage{
sampleH(
  nA,
  nB,
  links,
  survivalpSameH,
  pivs_stable,
  pivsA,
  pivsB,
  nvalues,
  nonlinkedA,
  nonlinkedB,
  eta,
  phi
)
}
\arguments{
\item{nA}{IntegerVector of dimensions of registered values of the Partially Identifying Variables (PIVs) in A}

\item{nB}{IntegerVector of dimensions of registered values of the PIVs in B}

\item{links}{IntegerMatrix of 2 columns with the indices of the linked records}

\item{survivalpSameH}{NumericMatrix with for each PIV the probability that true values coincide (if stable: filled with 1)}

\item{pivs_stable}{LogicalVector indicating for each PIV whether it is stable of not (if not we expect survivalpSameH for that same element to not be filled with 1 but with lower values)}

\item{pivsA}{List ith registered data from A}

\item{pivsB}{List with registered data from B}

\item{nvalues}{IntegerVector with number of unique values of each PIV}

\item{nonlinkedA}{LogicalVector indicating for all records in A whether they are linked or not}

\item{nonlinkedB}{LogicalVector indicating for all records in B whether they are linked or not}

\item{eta}{List parameters of the PIVs distributions}

\item{phi}{List parameters of the PIVs registration errors}
}
\value{
List:
\itemize{
\item truePIVsA, true values underlying data in A
\item truePIVsB, true values underlying data in B
}
}
\description{
sampleH
}
