% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleL}
\alias{sampleL}
\title{sampleNL}
\usage{
sampleL(
  GA,
  GB,
  survivalpSameH,
  choice_set,
  choice_equal,
  nval,
  phikA,
  phikB,
  eta
)
}
\arguments{
\item{GA}{IntegerVector of registered values for a certain Partially Identifying Variable (PIV) for linked records from A}

\item{GB}{IntegerVector of registered values for a certain PIV for linked records from B}

\item{survivalpSameH}{NumericVector of probabilities that the concerned PIV values coincide between file A and file B}

\item{choice_set}{IntegerMatrix of 2 columns (for A and for B) with possible joint true values underlying GA and GB}

\item{choice_equal}{IntegerVector of booleans indicating whether the 2 true values (from A and B) in the choice set are equal}

\item{nval}{integer for the number of unique values in the PIV concerned}

\item{phikA}{NumericVector parameter for the registration errors in A for the PIV concerned}

\item{phikB}{NumericVector parameter for the registration errors in B for the PIV concerned}

\item{eta}{NumericVector parameter for the distribution of the PIV concerned}
}
\value{
IntegerVector: of indices from the joint latent true values choice set underlying GA and GB
}
\description{
sampleNL
}
