% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleNL}
\alias{sampleNL}
\title{sampleNL}
\usage{
sampleNL(G, eta, phi)
}
\arguments{
\item{G}{IntegerVector of registered values for a certain Partially Identifying Variable (PIV) for non linked records}

\item{eta}{NumericVector parameter for the distribution of the PIV concerned}

\item{phi}{NumericVector parameter for the registration errors for the PIV concerned}
}
\value{
IntegerVector: of latent true values underlying G
}
\description{
sampleNL
}
