% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sspaste2}
\alias{sspaste2}
\title{sspaste2}
\usage{
sspaste2(A)
}
\arguments{
\item{A}{IntegerMatrix with values to form patterns}
}
\value{
CharacterVector: "sspaste2(A)" is a faster version in Rcpp of "do.call(paste, c(as.data.frame(A), list(sep="_")))"
}
\description{
sspaste2
}
