% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{summary.rl_diagnostics}
\alias{summary.rl_diagnostics}
\title{Summarises RL diagnostics}
\usage{
\method{summary}{rl_diagnostics}(object, ...)
}
\arguments{
\item{object}{An \code{rl_diagnostics} object.}

\item{...}{Extra arguments.}
}
\description{
Summarises RL diagnostics
}
