% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{compute_histogram_support_iou}
\alias{compute_histogram_support_iou}
\title{Intersection-over-union of two histogram supports}
\usage{
compute_histogram_support_iou(h1, h2)
}
\arguments{
\item{h1}{\code{histogram} object (as returned by \code{\link[graphics:hist]{graphics::hist()}}).}

\item{h2}{\code{histogram} object (as returned by \code{\link[graphics:hist]{graphics::hist()}}).}
}
\value{
Numeric, IoU of the ranges over which \code{h1} and \code{h2}
have non-zero counts.
}
\description{
Intersection-over-union of two histogram supports
}
\examples{
h1 <- graphics::hist(rnorm(200), plot = FALSE)
h2 <- graphics::hist(rnorm(200) + 1, plot = FALSE)
compute_histogram_support_iou(h1, h2)
}
