% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{createDataAlpha}
\alias{createDataAlpha}
\title{Title}
\usage{
createDataAlpha(nCoefUnstable, stable)
}
\arguments{
\item{nCoefUnstable}{.}

\item{stable}{.}
}
\value{
.
}
\description{
Title
}
\examples{
.
}
