% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{FDPinRL_fastLink}
\alias{FDPinRL_fastLink}
\title{FDP estimation with the fastLink package}
\usage{
FDPinRL_fastLink(NewA, NewB, arguments)
}
\arguments{
\item{NewA}{Data set as returned by \code{\link[=synthesise]{synthesise()}}.}

\item{NewB}{Data sets as returned by \code{\link[=synthesise]{synthesise()}}.}

\item{arguments}{List of extra arguments forwarded to \code{fastLink::fastLink()}
(e.g. \code{varnames}, \code{threshold.match}, \code{tol.em}, \code{return.all}).}
}
\value{
List with \code{result_index_linked_A}, \code{result_index_linked_B},
\code{result_vec_linked_scores}.
}
\description{
More details about fastLink on
https://cran.r-project.org/web/packages/fastLink/index.html
}
