% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{FDPinRL_reclin2}
\alias{FDPinRL_reclin2}
\title{FDP estimation with the reclin2 package}
\usage{
FDPinRL_reclin2(NewA, NewB, arguments)
}
\arguments{
\item{NewA, }{NewB Data sets as returned by \code{\link[=synthesise]{synthesise()}}.}

\item{arguments}{List of extra arguments forwarded across the \code{reclin2}
pipeline (\code{pair()}, \code{compare_pairs()}, \code{problink_em()}, \code{predict()},
\code{select_threshold()}): e.g. \code{on}, \code{formula}, \code{type}, \code{add}, \code{variable},
\code{score}, \code{threshold}.}
}
\value{
List with \code{result_index_linked_A}, \code{result_index_linked_B}, \code{result_vec_linked_scores}.
}
\description{
More details about reclin2 on https://cran.r-project.org/web/packages/reclin2/index.html
}
