% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{launchNaive}
\alias{launchNaive}
\title{Title}
\usage{
launchNaive(PIVs, encodedA, encodedB)
}
\arguments{
\item{PIVs}{.}

\item{encodedA}{.}

\item{encodedB}{.}
}
\value{
.
}
\description{
Title
}
\examples{
.
}
