% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{loglik}
\alias{loglik}
\title{Title}
\usage{
loglik(LLL, LLA, LLB, D, links, sumRowD, sumColD, gamma)
}
\arguments{
\item{LLL}{.}

\item{LLA}{.}

\item{LLB}{.}

\item{D}{.}

\item{links}{.}

\item{sumRowD}{.}

\item{sumColD}{.}

\item{gamma}{.}
}
\value{
.
}
\description{
Title
}
\examples{
.
}
