% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{loglikSurvival}
\alias{loglikSurvival}
\title{Title}
\usage{
loglikSurvival(alphas, X, times, Hequal)
}
\arguments{
\item{alphas}{A vector of size 1+cov in A+cov in B with coefficients of the hazard (baseline hazard and conditional hazard)}

\item{X}{A matrix with number of linked records rows and 1+cov in A+cov in B columns (first column: intercept, following columns: covariates from A and then from B to model instability) (used for optimisation: X concatenate the X obtained in each iteration of the Gibbs sampler)}

\item{times}{A vector of size number of linked records with the time gaps between the record from each sources (used for optimisation: times concatenate the times vectors obtained in each iteration of the Gibbs sampler)}

\item{Hequal}{A vector of size number of linked records with boolean values indicating wether the values in A and in B coincide (used for optimisation: times concatenate the times vectors obtained in each iteration of the Gibbs sampler)}
}
\value{
The value of the opposite (-) of the log likelihood associated with the survival function defining the probabilities that true values coincide (as defined in the paper) (the algorithm minimises -loglik i.e. maximises the loglik)
}
\description{
Title
}
\examples{
.
}
