% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{logPossibleConfig}
\alias{logPossibleConfig}
\title{Title}
\usage{
logPossibleConfig(Brecords, sumD)
}
\arguments{
\item{Brecords}{.}

\item{sumD}{.}
}
\value{
.
}
\description{
Title
}
\examples{
.
}
