% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{logPossibleConfig}
\alias{logPossibleConfig}
\title{logPossibleConfig}
\usage{
logPossibleConfig(Brecords, sumD)
}
\arguments{
\item{Brecords}{Number of records in data source B (the largest).}

\item{sumD}{Number of linked records (at a specific time point of the algorithm).}
}
\value{
The sum of logs of the vector going from nB - nLinks + 1 to nB.
}
\description{
This function helps calculating the number of possible designs for Delta given by nB!/(nB-nLinks)! Needed to compute the log likelihood of the linkage matrix.
}
\examples{
sumColD = c(0,0,0,0,1,0,0,0,1,0,1,0,1,0,1)
links = base::data.frame(list(idxA=c(5,9,11,12,13), idxB=c(5,9,11,13,15)))
possconfig = logPossibleConfig(length(sumColD),nrow(links))
}
