% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{print.rl_diagnostics}
\alias{print.rl_diagnostics}
\title{Display RL diagnostics}
\usage{
\method{print}{rl_diagnostics}(x, ...)
}
\arguments{
\item{x}{An \code{rl_diagnostics} object.}

\item{...}{Extra arguments.}
}
\value{
An object of class \code{"rl_diagnostics"}, a list with:
}
\description{
Display RL diagnostics
}
