% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{simulateD}
\alias{simulateD}
\title{Title}
\usage{
simulateD(
  data,
  D,
  linksR,
  sumRowD,
  sumColD,
  truepivsA,
  truepivsB,
  gamma,
  eta,
  alpha,
  phi
)
}
\arguments{
\item{data}{.}

\item{D}{.}

\item{linksR}{.}

\item{sumRowD}{.}

\item{sumColD}{.}

\item{truepivsA}{.}

\item{truepivsB}{.}

\item{gamma}{.}

\item{eta}{.}

\item{alpha}{.}

\item{phi}{.}
}
\value{
.
}
\description{
Title
}
\examples{
.
}
