% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{simulateH}
\alias{simulateH}
\title{Title}
\usage{
simulateH(data, D, links, survivalpSameH, sumRowD, sumColD, eta, phi)
}
\arguments{
\item{data}{.}

\item{D}{.}

\item{links}{.}

\item{survivalpSameH}{.}

\item{sumRowD}{.}

\item{sumColD}{.}

\item{eta}{.}

\item{phi}{.}
}
\value{
.
}
\description{
Title
}
\examples{
.
}
