% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{stEM}
\alias{stEM}
\title{Title}
\usage{
stEM(
  data,
  StEMIter,
  StEMBurnin,
  GibbsIter,
  GibbsBurnin,
  sparseOutput = TRUE,
  musicOn = TRUE,
  newDirectory = NULL,
  saveInfoIter = FALSE
)
}
\arguments{
\item{data}{A list with elements A, B, Nvalues, PIVs_config, controlOnMistakes,
sameMistakes, phiMistakesAFixed, phiMistakesBFixed, phiForMistakesA, phiForMistakesB}

\item{StEMIter}{An integer with the total number of iterations of the Stochastic
EM algorithm (including the period to discard as burn-in)}

\item{StEMBurnin}{An integer with the number of iterations to discard as burn-in}

\item{GibbsIter}{An integer with the total number of iterations of the Gibbs sampler
(done in each iteration of the StEM) (including the period to discard as burn-in)}

\item{GibbsBurnin}{An integer with the number of iterations to discard as burn-in}

\item{sparseOutput}{A boolean value, to handle the final linkage matrix as sparse or
not, for large data sets it should be set to TRUE to avoid R crashing or memory errors}

\item{musicOn}{A boolean value, if TRUE the algorithm will play music at the end of
the algorithm, useful if you have to wait for the record linkage to run and to act as
an alarm when record linkage is done}

\item{newDirectory}{A NULL value or: A string with the name of (or path to) the directory
(which should already exist) where to save the environment variables at the end of each
iteration (useful when record linkage is very long, to not loose everything and not restart
from scratch in case your computer shut downs before record linkage is finished)}

\item{saveInfoIter}{A boolean value to indicate whether you want the environment variables
to be saved at the end of each iteration (useful when record linkage is very long, to not
loose everything and not restart from scratch in case your computer shut downs before
record linkage is finished)}
}
\value{
A list with: Delta, (sparse) matrix with the pairs of records linked and their posterior probabilities to be linked (select the pairs where the proba>0.5 to get a valid set of linked records), gamma, a vector with the chain of the parameter gamma representing the proportion of linked records as a fraction of the smallest file, eta, a vector with the chain of the parameter eta representing the distribution of the PIVs, alpha, a vector with the chain of the parameter alpha representing the hazard coefficient of the model for instability, phi, a vector with the chain of the parameter phi representing the registration errors parameters)
}
\description{
Title
}
\examples{
data = list( A=encodedA,
                       B=encodedB,
                       Nvalues=nvalues,
                       PIVs_config=PIVs_config,
                       controlOnMistakes=c(TRUE,TRUE,FALSE,FALSE,FALSE),
                       sameMistakes=TRUE,
                       phiMistakesAFixed=FALSE,
                       phiMistakesBFixed=FALSE,
                       phiForMistakesA=c(NA,NA,NA,NA,NA),
                       phiForMistakesB=c(NA,NA,NA,NA,NA))
           fit = StEM( data,
                       100,
                       70,
                       200,
                       100,
                       TRUE,
                       TRUE,
                       NULL,
                       FALSE )
}
