% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRecordLinkage.r
\name{SurvivalUnstable}
\alias{SurvivalUnstable}
\title{Title}
\usage{
SurvivalUnstable(Xlinksk, alphask, times)
}
\arguments{
\item{Xlinksk}{A matrix with number of linked records rows and 1+cov in A+cov in B columns, with a first column filled with 1 (intercept), and following columns filled with the values of the covariates useful for modelling instability for the linked records}

\item{alphask}{A vector of size 1+cov in A+cov in B, with as first element the baseline hazard and following elements being the coefficient of the conditional hazard associated with the covariates given in X}

\item{times}{A vector of size number of linked records with the time gaps between the record from each sources}
}
\value{
A vector (for an unstable PIV) of size number of linked records with the probabilities that true values coincide (e.g. 1 - proba to move if the PIV is postal code) defined according to the survival function with exponential model proposed in the paper
}
\description{
Title
}
\examples{
.
}
